testenvironment/LFM2_5_VL_1_6B/requirements.txt
git+https://github.com/huggingface/transformers.git@2a5ba8b53d298ed8421e09831bf96bb6d056a24d
pillow
torch
accelerate
opencv-python

